+++ src/main/java/pl/peterwolf/cinewolf/CineWolfAutoDirector.java
@@ -3,7 +3,7 @@ package pl.peterwolf.cinewolf;
 public final class CineWolfAutoDirector {
     public static final String MOD_ID = "cinewolf_autodirector";
     public static final String MOD_NAME = "CineWolf AutoDirector";
-    public static final String VERSION = "2.0.0";
+    public static final String VERSION = "2.0.1";
 
     private CineWolfAutoDirector() {
     }